feat: move portable composables to commons (batch 20)
Move 11 files with zero Android-specific imports to commons module:

Components:
- GenericLoadable (sealed class)
- UrlPreviewState (sealed class)
- ForwardingPainter (painter utilities)

Toast hierarchy:
- ToastMsg, StringToastMsg, ActionableStringToastMsg
- ThrowableToastMsg, ThrowableToastMsg2
- ResourceToastMsg

Note creator interfaces:
- IExpiration, IMessageField, IZapRaiser
- DraftTagState
- SplitBuilder, SplitItem

Original files replaced with typealiases for backward compatibility.
Consumer imports updated to point directly to commons where needed.

Part of the KMP iOS migration tracked in #2238.
