+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/components/GenericLoadable.kt
@@ -20,19 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.components
 
-import androidx.compose.runtime.Immutable
-
-@Immutable
-sealed class GenericLoadable<T> {
-    @Immutable class Loading<T> : GenericLoadable<T>()
-
-    @Immutable class Loaded<T>(
-        val loaded: T,
-    ) : GenericLoadable<T>()
-
-    @Immutable class Empty<T> : GenericLoadable<T>()
-
-    @Immutable class Error<T>(
-        val errorMessage: String,
-    ) : GenericLoadable<T>()
-}
+// Re-export from commons for backwards compatibility
+typealias GenericLoadable<T> = com.vitorpamplona.amethyst.commons.ui.components.GenericLoadable<T>
